post page, user page added

+++ app/Http/Controllers/PostController.php
@@ -7,10 +7,20 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+    public function __construct() {
+        $this->middleware(['auth'])->only('store', 'destroy');
+    }
+
     public function index() {
         // Get all the post to output to screen
         $posts = Post::with(['user', 'likes'])->orderBy('created_at', 'desc')->paginate(5);
         return view('posts.index')->with('posts', $posts);
+    }
+
+    public function show(Post $post) {
+        return view('posts.show', [
+            'post' => $post
+        ]);
     }
 
     public function store(Request $request) {
@@ -28,9 +38,8 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post) {
-        if ($post->ownedBy(auth()->user())) {
-            //
-        }
+        
+        $this->authorize('delete', $post);
 
         $post->delete();
